Factor out name mangling into separate class
